Discriminant -> double

+++ include/equation.h
@@ -6,7 +6,7 @@
 /*   By: fhuang <fhuang@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/12/06 16:50:12 by fhuang            #+#    #+#             */
-/*   Updated: 2017/12/18 16:48:38 by fhuang           ###   ########.fr       */
+/*   Updated: 2017/12/18 19:18:18 by fhuang           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ class	equation
 		std::string		x1;
 		std::string		x2;
 		bool			debug;
-		long			discriminant;
+		double			discriminant;
 		int				degree;
 
 	private:
